[DAG] insert_data_to_db DAG 추가 (#5)
* [dag] DAG 파일 구조 작성

* [dag] dto 필드 변경 및 table 추가

* [dag] dag 완성 및 import 에러 수정

+++ dto/air_quality.py
@@ -1,7 +1,5 @@
 import datetime
-from dataclasses import Field
-
-from pydantic import BaseModel, validator
+from pydantic import BaseModel, validator, Field
 
 
 class AirQualityDTO(BaseModel):
@@ -11,8 +9,8 @@ class AirQualityDTO(BaseModel):
     o3: float
     co: float
     so2: float
-    pm10: int
-    pm25: int
+    pm10: float
+    pm25: float
 
     @validator("measure_datetime")
     def string_to_datetime(cls, value: str) -> datetime.datetime:
